[e2e] Add pandas join test
This patch adds an example of performing a `join` operation using
`pd.DataFrame`s. To run this test:

```shell
$ cd python/test
$ pytest -s -k 'test_convert[../../e2e_testing/join.py]'
```

NOTE: This test currently passes when it shouldn't. The exit code is
0, but we output the error(s) below. As a result, this test is marked
as an `xfail` because I don't *think* this should work out of the box yet.

```text
test_ast.py error: empty block: expect at least a terminator
error: failed to run passes
```

We also add an explicit `cmake_configure.sh` script and update the
docs to make getting started a little easier.